allow ssl and non-ssl traffic to connect to pg
When we connect to the database using loopback, a password
is not checked. So these configurations did not notice ssl requirement

More and more people are trying to connect to the postgres server
without ssl. This is causing them to fail.

This configuration change allows clients to connect to the database
without configuring pg ssl certificates and encryption